UIDATIMP-633 Relocate the file upload area on the Data Import landing page
