+++ src/components/ImportJobs/ImportJobs.css
@@ -5,11 +5,16 @@
   align-items: center;
   flex-direction: column;
   justify-content: center;
-  height: 100%;
+  height: 150px;
   padding: var(--gutter);
+  margin-bottom: var(--gutter);
   border: 1px dashed var(--color-border);
   border-radius: var(--radius);
   background-color: var(--color-fill);
+
+& button[class*="button"] {
+    margin: 0;
+  }
 }
 
 .uploadTitle {
