--- src/components/ImportJobs/components/FileUploader/defaultStyles.css
@@ -1,12 +1,0 @@
-@import "@folio/stripes-components/lib/variables.css";
-
-.defaultFileUploader {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  min-height: 200px;
-  border: 1px dashed var(--color-border);
-  background-color: var(--color-fill);
-  border-radius: var(--radius);
-}
