--- src/components/ImportJobs/components/FileUploader/index.js
@@ -1,1 +1,0 @@
-export * from './FileUploader';
