+++ src/components/ImportJobs/components/index.js
@@ -1,2 +1,1 @@
-export * from './FileUploader';
 export * from './ReturnToAssignJobs';
