+++ translations/ui-data-import/en.json
@@ -2,7 +2,6 @@
   "meta.title": "Data Import",
   "jobsPaneTitle": "Jobs",
   "logsPaneTitle": "Logs",
-  "importPaneTitle": "Import",
   "viewAllLogs": "View all",
   "manageJobs": "Manage",
   "previewJobs": "Previews",
